demo/requirements.txt
autopep8==1.5.7
grpcio==1.39.0
grpcio-tools==1.39.0
kafka-python==2.0.2
numpy==1.21.1
Pillow==8.3.1
protobuf==3.17.3
pycodestyle==2.7.0
six==1.16.0
toml==0.10.2
